mito-ai: rename VariableManager to ContextManager in more places
