+++ mito-ai/src/Extensions/AiChat/ChatHistoryManager.tsx
@@ -1,5 +1,5 @@
 import OpenAI from "openai";
-import { IContextManager } from "../VariableManager/VariableManagerPlugin";
+import { IContextManager } from "../ContextManager/ContextManagerPlugin";
 import { INotebookTracker } from '@jupyterlab/notebook';
 import { getActiveCellCode, getActiveCellID, getCellCodeByID } from "../../utils/notebook";
 import { IAgentPlanningMetadata, IChatMessageMetadata, ICodeExplainMetadata, ISmartDebugMetadata } from "../../utils/websocket/models";
