+++ mito-ai/src/Extensions/ContextManager/FileInspector.tsx
