+++ mito-ai/src/Extensions/ContextManager/VariableInspector.tsx
